+++ crates/storybook/src/storybook.rs
@@ -3,9 +3,8 @@
 mod collab_panel;
 mod stories;
 mod story;
+mod story_selector;
 mod workspace;
-
-use std::str::FromStr;
 
 use ::theme as legacy_theme;
 use clap::Parser;
@@ -19,61 +18,19 @@ use stories::components::facepile::FacepileStory;
 use stories::components::toolbar::ToolbarStory;
 use stories::components::traffic_lights::TrafficLightsStory;
 use stories::elements::avatar::AvatarStory;
-use strum::EnumString;
 use ui::{ElementExt, Theme};
+
+use crate::story_selector::{ComponentStory, ElementStory, StorySelector};
 
 gpui2::actions! {
     storybook,
     [ToggleInspector]
 }
 
-#[derive(Debug, Clone, Copy)]
-enum StorySelector {
-    Element(ElementStory),
-    Component(ComponentStory),
-}
-
-impl FromStr for StorySelector {
-    type Err = anyhow::Error;
-
-    fn from_str(raw_story_name: &str) -> std::result::Result<Self, Self::Err> {
-        let story = raw_story_name.to_ascii_lowercase();
-
-        if let Some((_, story)) = story.split_once("elements/") {
-            let element_story = ElementStory::from_str(story)
-                .with_context(|| format!("story not found for element '{story}'"))?;
-
-            return Ok(Self::Element(element_story));
-        }
-
-        if let Some((_, story)) = story.split_once("components/") {
-            let component_story = ComponentStory::from_str(story)
-                .with_context(|| format!("story not found for component '{story}'"))?;
-
-            return Ok(Self::Component(component_story));
-        }
-
-        Err(anyhow!("story not found for '{raw_story_name}'"))
-    }
-}
-
-#[derive(Debug, Clone, Copy, EnumString)]
-#[strum(serialize_all = "snake_case")]
-enum ElementStory {
-    Avatar,
-}
-
-#[derive(Debug, Clone, Copy, EnumString)]
-#[strum(serialize_all = "snake_case")]
-enum ComponentStory {
-    Breadcrumb,
-    Facepile,
-    Toolbar,
-    TrafficLights,
-}
-
 #[derive(Parser)]
+#[command(author, version, about, long_about = None)]
 struct Args {
+    #[arg(value_enum)]
     story: Option<StorySelector>,
 }
 
@@ -146,7 +103,7 @@ fn current_theme<V: 'static>(cx: &mut ViewContext<V>) -> Theme {
         .clone()
 }
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{anyhow, Result};
 use gpui2::AssetSource;
 use rust_embed::RustEmbed;
 use workspace::WorkspaceElement;
